Updated inflation rate function
The inflation function now follows a gaussian distribution, instead of a
uniform one. Also, the average rate was calculated linearly. It is now
calculated as root 365 of the annual rate, as it should.

The num crate and related helper function were removed as well, since
f64 offers enough precision for our purposes.